Add a configuration typer application ...
Creates a commands module to store the command logic in one location.
Moves much settings configuration into the config submodule in commands
Attempts to create a clear logically distinction between cli logic and
hub logic
